requirements.txt
numpy>=1.24
requests>=2.28
websocket-client>=1.6
